pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hermes-telemetry"
version = "0.1.0"
description = "OpenTelemetry observability plugin for Hermes Agent"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "jzb"},
]
dependencies = [
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
]

[project.entry-points."hermes_agent.plugins"]
hermes-telemetry = "hermes_telemetry:register"

[tool.setuptools.packages.find]
include = ["hermes_otel*"]
